server/buck: Add buck support to build event handler
This change adds support for processing Buck build events alongside
Bazel events.  A new `buck` package is introduced to handle the
conversion of Buck-specific events into a format compatible with the
existing build event processing pipeline. The build event handler is
modified to utilize this new package when encountering a Buck event.